+++ src/main/java/com/ToDoList2/ToDoList2/repository/ToDoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ToDoList2.ToDoList2.entity.ToDo;
 
-public interface ToDoRepository extends JpaRepository<ToDo, Long> {
+public interface ToDoRepository extends JpaRepository<ToDo, Integer> {
     
 }
