+++ src/main/java/com/ToDoList2/ToDoList2/service/impl/ToDoServiceImpl.java
@@ -27,7 +27,7 @@ public class ToDoServiceImpl implements ToDoService {
     }
 
     @Override
-    public ToDo getToDoById(Long id) {
+    public ToDo getToDoById(Integer id) {
         Optional<ToDo> toDo = this.toDoRepository.findById(id);
         if (!toDo.isPresent()) {
             throw new ResourceNotFoundException("ToDo not found with id: " + id);
@@ -42,7 +42,7 @@ public class ToDoServiceImpl implements ToDoService {
     }
 
     @Override
-    public ToDo updateToDo(Long id, ToDo toDo) {
+    public ToDo updateToDo(Integer id, ToDo toDo) {
         Optional<ToDo> toDoToUpdateOptional = this.toDoRepository.findById(id);
         if(!toDoToUpdateOptional.isPresent()) {
             throw new ResourceNotFoundException("ToDo not found with id: " + id);
@@ -67,7 +67,7 @@ public class ToDoServiceImpl implements ToDoService {
     }
 
         @Override
-    public ToDo deleteToDo(Long id) {
+    public ToDo deleteToDo(Integer id) {
         Optional<ToDo> toDoToDeleteOptional = this.toDoRepository.findById(id);
 
         if (!toDoToDeleteOptional.isPresent()) {
